requirements.txt
flask
pandas
requests
python-dotenv
gdown
scikit-learn
numpy